+++ CHANGELOG.md
@@ -21,3 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated `mix set-output` to remove unassigned devices from the mix instead of reassigning them to another mix.
 - Removed the requirement for at least two mixes when using `mix set-output`.
 
+
+## [0.0.4] - 2026-01-11
+
+### Added
+- Added `channel toggle-mute-in-mix` command.
+- Added `channel isolate` command to mute all channels in a mix except one.